Show last positiosn when showing vessel beacon malfunction with resumed transmission
